# internal/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: SendRec API
  version: "1.57.1"
  description: API for SendRec, an open-source EU-native async video messaging platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://app.sendrec.eu
    description: Production
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 72
    ConfirmEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ResendConfirmationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    UserResponse:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Rate limited to 0.5 requests per second with a burst of 5. Returns 403 when registration is disabled via `REGISTRATION_ENABLED=false`.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created. A confirmation email is sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Registration is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: Rate limited to 0.5 requests per second with a burst of 5.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 'Email not verified. Error body contains {"error": "email_not_verified"}.'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Uses the refresh_token cookie to issue a new access token. Rate limited to 0.5 requests per second with a burst of 5.
      operationId: refreshToken
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Always returns 200 to prevent email enumeration. Rate limited to 0.5 requests per second with a burst of 5.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with token
      description: Rate limited to 0.5 requests per second with a burst of 5.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token, or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/confirm-email:
    post:
      tags: [Authentication]
      summary: Confirm email address
      description: Verifies the confirmation token from the email link and activates the account.
      operationId: confirmEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmEmailRequest"
      responses:
        "200":
          description: Email confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/resend-confirmation:
    post:
      tags: [Authentication]
      summary: Resend confirmation email
      description: Always returns 200 to prevent email enumeration. If the email exists and is unverified, a new confirmation email is sent. Token expires in 24 hours.
      operationId: resendConfirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendConfirmationRequest"
      responses:
        "200":
          description: Confirmation email sent if account exists and is unverified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/user:
    get:
      tags: [User]
      summary: Get current user profile
      description: Rate limited to 2 requests per second with a burst of 10.
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [User]
      summary: Update current user profile
      description: Rate limited to 2 requests per second with a burst of 10.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong current password or unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/notifications:
    get:
      tags: [Settings]
      summary: Get notification preferences
      description: Returns the current user's account-wide notification preferences.
      operationId: getNotificationPreferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Settings]
      summary: Update notification preferences
      description: Sets the account-wide notification preference for view and comment notifications.
      operationId: updateNotificationPreferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesRequest"
      responses:
        "204":
          description: Preferences updated
        "400":
          description: Invalid notification mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/notifications/test-slack:
    post:
      tags: [Settings]
      summary: Send test Slack notification
      description: Sends a test message to the user's configured Slack webhook URL. The webhook URL must be saved first via PUT /api/settings/notifications.
      operationId: testSlackWebhook
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Test message sent successfully
        "400":
          description: No Slack webhook URL configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Failed to send message to Slack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/branding:
    get:
      tags: [Settings]
      summary: Get branding settings
      description: Returns the current user's watch page branding customization. Null fields use SendRec defaults. Requires branding to be enabled.
      operationId: getBrandingSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Branding settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingSettingsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Branding not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Settings]
      summary: Update branding settings
      description: Sets the account-wide watch page branding. Null or empty fields are stored as NULL (use SendRec defaults). Requires branding to be enabled.
      operationId: updateBrandingSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBrandingRequest"
      responses:
        "204":
          description: Branding updated
        "400":
          description: Validation error (invalid color, name too long)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Branding not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/branding/logo:
    post:
      tags: [Settings]
      summary: Upload branding logo
      description: Returns a presigned S3 URL for uploading a custom logo. PNG or SVG, max 512KB. Requires branding to be enabled.
      operationId: uploadBrandingLogo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoUploadRequest"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoUploadResponse"
        "400":
          description: Invalid content type or size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Branding not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Settings]
      summary: Delete branding logo
      description: Removes the custom logo from S3 and clears the logo key. Requires branding to be enabled.
      operationId: deleteBrandingLogo
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logo deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Branding not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/api-keys:
    get:
      tags: [Settings]
      summary: List API keys
      description: Returns all API keys for the authenticated user. Keys are not returned in plaintext.
      operationId: listAPIKeys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKeyItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Settings]
      summary: Generate API key
      description: |
        Creates a new API key for the authenticated user. The plaintext key is returned only once in the response.
        Maximum 10 keys per user. Keys use the `sr_` prefix.
      operationId: generateAPIKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  description: Optional label for the key (e.g. "My Nextcloud")
      responses:
        "201":
          description: API key created (plaintext key returned only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAPIKeyResponse"
        "400":
          description: Validation error or key limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/api-keys/{id}:
    delete:
      tags: [Settings]
      summary: Delete API key
      operationId: deleteAPIKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: API key deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...